Generate heatmap
The script iterates a list of files and outputs the heatmap of all
routers for each file in each one line. The files are either given as
parameter or all csv files in the current folder are used if no
parameter is given.

The heatmap is a dictionary with router ids and the mean service time
at all buffers in the routers.